requirements.txt
streamlit>=1.38.0
openai>=1.44.0
python-dotenv
